2/6/2024
Created basic prefabs for World and UI objects that can be clicked on and interacted with. World objects can have a UI object linked to them that will open if the world obj is clicked on. Else it will perform its own interaction behavior. UI interactables share the same behavior, but are in the UI instead of the world.